refactor(ga): separate helper function into utils
